requirements.txt
# requirements.txt

Flask
gunicorn
python-dotenv
requests
pytz
psycopg2-binary
